Moved Grimoire of Sacrifice module from Affliction to Shared
