+++ src/parser/warlock/shared/modules/talents/GrimoireOfSacrifice.js
