+++ src/parser/warlock/shared/modules/talents/normalizers/GrimoireOfSacrificeNormalizer.js
